Rename python package to neutron_plugin_contrail
